fix: register macos file-association listener once in app root
useFileOperations is consumed by EditorView, TabBar, and WelcomeScreen.
The init useEffect inside that hook ran per caller, so three 'open-file'
listeners were attached; an Open With event on an already-running app
fired all three listeners, opening three tabs for the same file.

Move the init (get_opened_file + listen, plus web-mode IndexedDB
restore) into App.tsx where it mounts exactly once, and clean up the
listener on unmount.
